Add files via upload
this is a boot screen to run on MacOS terminal and the sound that plays auto. If you want it to be active when you log in then go to system settings and go to users and groups then select the user and go in login elements and select + then select the applescript, so that when you log in the terminal opens , does the sound and shows you the logo.